refacto: Se Dividieron cada una de las historias de usuario en el paquete de eventos
