+++ src/main/java/eventos/HU022C3VerEventosRegistrados.java
@@ -1,0 +1,36 @@
+package eventos;
+
+public class HU022C3VerEventosRegistrados {
+    public class Eventos {
+        private String nombre;
+        private String fecha;
+        private String tipo;
+
+        public Eventos(String nombre, String fecha, String tipo) {
+            this.nombre = nombre;
+            this.fecha = fecha;
+            this.tipo = tipo;
+        }
+
+        public String getNombre() {
+            return nombre;
+        }
+
+        public String getFecha() {
+            return fecha;
+        }
+
+        public String getTipo() {
+            return tipo;
+        }
+
+        @Override
+        public String toString() {
+            return "Evento{" +
+                    "nombre='" + nombre + '\'' +
+                    ", fecha='" + fecha + '\'' +
+                    ", tipo='" + tipo + '\'' +
+                    '}';
+        }
+    }
+}
